Feat: create data module & move entity [#14]
