+++ data/src/main/java/com/sulgorae/data/entity/accounts/Accounts.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.accounts
+
+class Accounts : ArrayList<AccountsItem>()
