+++ data/src/main/java/com/sulgorae/data/entity/candles/days/Days.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.candles.days
+
+class Days : ArrayList<DaysItem>()
