+++ data/src/main/java/com/sulgorae/data/entity/market/all/All.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.market.all
+
+class All : ArrayList<AllItem>()
