+++ data/src/main/java/com/sulgorae/data/entity/market/all/AllItem.kt
@@ -1,0 +1,7 @@
+package com.sulgorae.data.entity.market.all
+
+data class AllItem(
+    val english_name: String,
+    val korean_name: String,
+    val market: String
+)
