+++ data/src/main/java/com/sulgorae/data/entity/order/Trade.kt
@@ -1,0 +1,10 @@
+package com.sulgorae.data.entity.order
+
+data class Trade(
+    val funds: String,
+    val market: String,
+    val price: String,
+    val side: String,
+    val uuid: String,
+    val volume: String
+)
