+++ data/src/main/java/com/sulgorae/data/entity/orderbook/Orderbook.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.orderbook
+
+class Orderbook : ArrayList<OrderbookItem>()
