+++ data/src/main/java/com/sulgorae/data/entity/orderbook/OrderbookItem.kt
@@ -1,0 +1,9 @@
+package com.sulgorae.data.entity.orderbook
+
+data class OrderbookItem(
+    val market: String,
+    val orderbook_units: List<OrderbookUnit>,
+    val timestamp: Long,
+    val total_ask_size: Double,
+    val total_bid_size: Double
+)
