+++ data/src/main/java/com/sulgorae/data/entity/orders/Orders.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.orders
+
+class Orders : ArrayList<OrdersItem>()
