+++ data/src/main/java/com/sulgorae/data/entity/orders/chance/Bid.kt
@@ -1,0 +1,7 @@
+package com.sulgorae.data.entity.orders.chance
+
+data class Bid(
+    val currency: String,
+    val min_total: Int,
+    val price_unit: Any
+)
