+++ data/src/main/java/com/sulgorae/data/entity/orders/chance/Chance.kt
@@ -1,0 +1,9 @@
+package com.sulgorae.data.entity.orders.chance
+
+data class Chance(
+    val ask_account: AskAccount,
+    val ask_fee: String,
+    val bid_account: BidAccount,
+    val bid_fee: String,
+    val market: Market
+)
