+++ data/src/main/java/com/sulgorae/data/entity/ticker/Ticker.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.ticker
+
+class Ticker : ArrayList<TickerItem>()
