+++ data/src/main/java/com/sulgorae/data/entity/trades/ticks/Ticks.kt
@@ -1,0 +1,3 @@
+package com.sulgorae.data.entity.trades.ticks
+
+class Ticks : ArrayList<TicksItem>()
