+++ data/src/main/java/com/sulgorae/data/entity/trades/ticks/TicksItem.kt
@@ -1,0 +1,13 @@
+package com.sulgorae.data.entity.trades.ticks
+
+data class TicksItem(
+    val ask_bid: String,
+    val chane_price: Int,
+    val market: String,
+    val prev_closing_price: Int,
+    val timestamp: Long,
+    val trade_date_utc: String,
+    val trade_price: Int,
+    val trade_time_utc: String,
+    val trade_volume: Double
+)
