--- domain/src/main/java/com/sulgorae/domain/entity/accounts/Accounts.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.accounts
-
-class Accounts : ArrayList<AccountsItem>()
