--- domain/src/main/java/com/sulgorae/domain/entity/candles/days/Days.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.candles.days
-
-class Days : ArrayList<DaysItem>()
