--- domain/src/main/java/com/sulgorae/domain/entity/market/all/All.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.market.all
-
-class All : ArrayList<AllItem>()
