--- domain/src/main/java/com/sulgorae/domain/entity/market/all/AllItem.kt
@@ -1,7 +1,0 @@
-package com.sulgorae.domain.entity.market.all
-
-data class AllItem(
-    val english_name: String,
-    val korean_name: String,
-    val market: String
-)
