--- domain/src/main/java/com/sulgorae/domain/entity/order/Trade.kt
@@ -1,10 +1,0 @@
-package com.sulgorae.domain.entity.order
-
-data class Trade(
-    val funds: String,
-    val market: String,
-    val price: String,
-    val side: String,
-    val uuid: String,
-    val volume: String
-)
