--- domain/src/main/java/com/sulgorae/domain/entity/orderbook/Orderbook.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.orderbook
-
-class Orderbook : ArrayList<OrderbookItem>()
