--- domain/src/main/java/com/sulgorae/domain/entity/orderbook/OrderbookItem.kt
@@ -1,9 +1,0 @@
-package com.sulgorae.domain.entity.orderbook
-
-data class OrderbookItem(
-    val market: String,
-    val orderbook_units: List<OrderbookUnit>,
-    val timestamp: Long,
-    val total_ask_size: Double,
-    val total_bid_size: Double
-)
