--- domain/src/main/java/com/sulgorae/domain/entity/orderbook/OrderbookUnit.kt
@@ -1,8 +1,0 @@
-package com.sulgorae.domain.entity.orderbook
-
-data class OrderbookUnit(
-    val ask_price: Int,
-    val ask_size: Double,
-    val bid_price: Int,
-    val bid_size: Double
-)
