--- domain/src/main/java/com/sulgorae/domain/entity/orders/Orders.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.orders
-
-class Orders : ArrayList<OrdersItem>()
