--- domain/src/main/java/com/sulgorae/domain/entity/orders/chance/Ask.kt
@@ -1,7 +1,0 @@
-package com.sulgorae.domain.entity.orders.chance
-
-data class Ask(
-    val currency: String,
-    val min_total: Int,
-    val price_unit: Any
-)
