--- domain/src/main/java/com/sulgorae/domain/entity/orders/chance/AskAccount.kt
@@ -1,10 +1,0 @@
-package com.sulgorae.domain.entity.orders.chance
-
-data class AskAccount(
-    val avg_buy_price: String,
-    val avg_buy_price_modified: Boolean,
-    val balance: String,
-    val currency: String,
-    val locked: String,
-    val unit_currency: String
-)
