--- domain/src/main/java/com/sulgorae/domain/entity/orders/chance/Chance.kt
@@ -1,9 +1,0 @@
-package com.sulgorae.domain.entity.orders.chance
-
-data class Chance(
-    val ask_account: AskAccount,
-    val ask_fee: String,
-    val bid_account: BidAccount,
-    val bid_fee: String,
-    val market: Market
-)
