--- domain/src/main/java/com/sulgorae/domain/entity/ticker/Ticker.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.ticker
-
-class Ticker : ArrayList<TickerItem>()
