--- domain/src/main/java/com/sulgorae/domain/entity/trades/ticks/Ticks.kt
@@ -1,3 +1,0 @@
-package com.sulgorae.domain.entity.trades.ticks
-
-class Ticks : ArrayList<TicksItem>()
