+++ settings.gradle
@@ -16,3 +16,4 @@ rootProject.name = "crypto"
 include ':app'
 include ':domain'
 include ':remote'
+include ':data'
